+++ db/sqlc/account.sql.go
@@ -39,10 +39,16 @@ func (q *Queries) CreateAccount(ctx context.Context, arg CreateAccountParams) (A
 const deleteAccount = `-- name: DeleteAccount :exec
 DELETE FROM accounts
 WHERE id = $1
+AND owner = $2
 `
 
-func (q *Queries) DeleteAccount(ctx context.Context, id int64) error {
-	_, err := q.db.ExecContext(ctx, deleteAccount, id)
+type DeleteAccountParams struct {
+	ID    int64  `json:"id"`
+	Owner string `json:"owner"`
+}
+
+func (q *Queries) DeleteAccount(ctx context.Context, arg DeleteAccountParams) error {
+	_, err := q.db.ExecContext(ctx, deleteAccount, arg.ID, arg.Owner)
 	return err
 }
 
@@ -86,17 +92,19 @@ func (q *Queries) GetAccountForUpdate(ctx context.Context, id int64) (Account, e
 const listAccounts = `-- name: ListAccounts :many
 SELECT id, owner, balance, currency, created_at FROM accounts
 WHERE id > $1
+AND owner = $2
 ORDER BY id
-LIMIT $2
+LIMIT $3
 `
 
 type ListAccountsParams struct {
-	ID    int64 `json:"id"`
-	Limit int32 `json:"limit"`
+	ID    int64  `json:"id"`
+	Owner string `json:"owner"`
+	Limit int32  `json:"limit"`
 }
 
 func (q *Queries) ListAccounts(ctx context.Context, arg ListAccountsParams) ([]Account, error) {
-	rows, err := q.db.QueryContext(ctx, listAccounts, arg.ID, arg.Limit)
+	rows, err := q.db.QueryContext(ctx, listAccounts, arg.ID, arg.Owner, arg.Limit)
 	if err != nil {
 		return nil, err
 	}
@@ -128,16 +136,18 @@ const updateAccount = `-- name: UpdateAccount :one
 UPDATE accounts
 SET balance = $2
 WHERE id = $1
+AND owner = $3
 RETURNING id, owner, balance, currency, created_at
 `
 
 type UpdateAccountParams struct {
 	ID      int64   `json:"id"`
 	Balance float64 `json:"balance"`
+	Owner   string  `json:"owner"`
 }
 
 func (q *Queries) UpdateAccount(ctx context.Context, arg UpdateAccountParams) (Account, error) {
-	row := q.db.QueryRowContext(ctx, updateAccount, arg.ID, arg.Balance)
+	row := q.db.QueryRowContext(ctx, updateAccount, arg.ID, arg.Balance, arg.Owner)
 	var i Account
 	err := row.Scan(
 		&i.ID,
